Move rule model to model::rule_model
This is a first step towards #86 and #227.
