+++ src/logical/model/chase_model.rs
@@ -1,0 +1,1 @@
+//! A variant of the rule model suitable for computing the chase.
